refactor: 공지 dto 패키지 변경
