+++ src/main/java/shop/tryit/domain/notice/dto/NoticeUpdateFormDto.java
@@ -1,20 +1,26 @@
-package shop.tryit.web.notice;
+package shop.tryit.domain.notice.dto;
 
+import javax.validation.constraints.NotEmpty;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
-public class NoticeViewFormDto {
+public class NoticeUpdateFormDto {
 
     private Long id;
+
+    @NotEmpty(message = "제목을 입력해주세요.")
     private String title;
+
+    @NotEmpty(message = "내용을 입력해주세요.")
     private String content;
+
     private String memberEmail;
 
     @Builder
-    private NoticeViewFormDto(Long id, String title, String content, String memberEmail) {
+    private NoticeUpdateFormDto(Long id, String title, String content, String memberEmail) {
         this.id = id;
         this.title = title;
         this.content = content;
